requirements.txt
# Python dependencies for Trustpilot scraper tools
# Install: pip install -r requirements.txt
# After install: python -m playwright install chromium

playwright>=1.40.0
python-dotenv>=1.0.0
postgrest>=0.13.0
playwright-stealth>=1.0.6
